requirements.txt
# requirements.txt
python-dotenv
requests
sentence-transformers
faiss-cpu
scikit-learn
numpy
playwright
tqdm